feat(forum_member): renaming www app
