+++ config/settings/base.py
@@ -72,6 +72,7 @@ LOCAL_APPS = [
     "lacommunaute.inclusion_connect",
     "lacommunaute.www.forum_views",
     "lacommunaute.www.forum_conversation_views",
+    "lacommunaute.www.forum_member_views",
 ]
 
 INSTALLED_APPS = DJANGO_APPS + LOCAL_APPS + THIRD_PARTIES_APPS
