+++ config/urls.py
@@ -6,6 +6,7 @@ from machina.core.loading import get_class
 
 from lacommunaute.www.forum_conversation_views import urls as forum_conversation_extension_urls
 from lacommunaute.www.forum_conversation_views.forum_polls import urls as forum_polls_extension_urls
+from lacommunaute.www.forum_member_views import urls as forum_member_urls
 from lacommunaute.www.forum_views import urls as forum_extension_urls
