+++ lacommunaute/www/forum_member_views/views.py
@@ -17,7 +17,7 @@ from machina.core.loading import get_class
 from lacommunaute.forum.models import Forum
 from lacommunaute.forum_member.models import ForumProfile
 from lacommunaute.utils.urls import get_safe_url
-from lacommunaute.www.forum_member.forms import JoinForumForm
+from lacommunaute.www.forum_member_views.forms import JoinForumForm
 
 
 logger = logging.getLogger(__name__)
